+++ subcommand/injector/command.go
@@ -40,6 +40,7 @@ type Command struct {
 	flagAutoName           string // MutatingWebhookConfiguration for updating
 	flagAutoHosts          string // SANs for the auto-generated TLS cert.
 	flagVaultService       string // Name of the Vault service
+	flagProxyAddress       string // HTTP proxy address used to talk to the Vault service
 	flagVaultImage         string // Name of the Vault Image to use
 	flagVaultAuthType      string // Type of Vault Auth Method to use
 	flagVaultAuthPath      string // Mount path of the Vault Auth Method
@@ -146,6 +147,7 @@ func (c *Command) Run(args []string) int {
 		VaultAddress:       c.flagVaultService,
 		VaultAuthType:      c.flagVaultAuthType,
 		VaultAuthPath:      c.flagVaultAuthPath,
+		ProxyAddress:       c.flagProxyAddress,
 		ImageVault:         c.flagVaultImage,
 		Clientset:          clientset,
 		RequireAnnotation:  true,